requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pandas==2.2.2
python-jose==3.3.0
python-dotenv==1.0.1
